AI: new module ShipDesignAI
The new module provides an extensive and flexible framework for the
design process of ships for the AI.

Currently ship classes used by the AI with the exception of scouts are
implemented. Future kind of special purpose ships can quickly be added
as most of the implementation is very generic, so only specifications
for the new shipclass need to be provided.